+++ js/components/braveryPanel.js
@@ -8,7 +8,7 @@ const ImmutableComponent = require('../../app/renderer/components/immutableCompo
 const config = require('../constants/config')
 const Dialog = require('./dialog')
 const SwitchControl = require('./switchControl')
-const {FormDropdown} = require('../../app/renderer/components/dropdown')
+const {FormDropdown} = require('../../app/renderer/components/common/dropdown')
 const windowActions = require('../actions/windowActions')
 const appActions = require('../actions/appActions')
 const urlParse = require('../../app/common/urlParse')
